PB-3478: Squashed commits to clean branch. This branch modifies mainly RTC module and requires the onnxruntime-node module. In particular, the localJitsiTrack.js and remoteJitsiTrack.js are modified to use the onnx models and modify the videos. The current lib-jitsi-meet has been developed with the jitsi's repo jitsi-meet. Since Internxt's meet-web does not link properly with lib-jitsi (at least on my computer). Further comments and notes are added to the PR.
PB-3478: Added somo in initial test using a .spec file. There is not a proper mock class for the JitsiLocalTrack, so it must be done to do more relevant tests

PB-3478: modified karma.conf so karma can access the onnx and wasm modules. I am still defining the tests and what should be their results, also, need to create a mock for participant

PB-3478: Added timers to check latency bottlenecks

PB-3478: cahnged date for console.time, unique ids were given to the timers

PB-3478: test for encoding at localTrack are currently successful. Nontheless, critical scenarios must be checked to see the behaviour of the encoder. Next tests should be related to the JVB

PB-3478: test for decoding added successfully

PB-3478: added a mock for JitsiLocalTrack in the tests, the mock needs to be tested (it actually mocks a JLT)

PB-3478: added a mocked jitsilocaltrack to the  corresponding .spec file and some tests on it. Performance tests must be done and edge ones aswell. JitsiRemoteTRack requires similar tests!

PB-3478: Added a mock for JitsiRemoteTrack. Initial tests have been added. Encoder has been disabled since Jitsi encoders reduce the size as expected by the decoder, it is not required to do it twice

PB-3478: Removed comments and added a flag to allow them to check models performance

PB-3478: Added a flag to the decoder so the decoder it is deactivated when the incoming stream is attached to the UI

PB-3478: hotfix to enable karma tests to run properly. Also increased the maximum resolution received and sent

PB-3478: Removed comments and fixed karma.conf to read the encoder and decoder files